Added an improved level and zombies

+++ 2dGame/src/entities/Mob.java
@@ -6,14 +6,14 @@ import game.level.tiles.Tile;
 public abstract class Mob extends Entity {
 	
 	protected String name;
-	protected int speed;
+	protected double speed;
 	protected int numSteps = 0;
 	protected boolean isMoving;
 	protected int movingDir = 1;
 	protected int scale = 1;
 	
 
-	public Mob(Level level, String name, int x, int y, int speed) {
+	public Mob(Level level, String name, int x, int y, double speed) {
 		super(level);
 		
 		this.name = name;
